Removing Blowout and Fogbow Manager Dependencies
